+++ tests/testthat/test-connect.R
@@ -24,3 +24,12 @@ test_that("Connecting to lamindata works", {
 
   expect_s3_class(artifact$wells, "RelatedRecords") # one-to-many
 })
+
+test_that("Connecting to a private instance works", {
+  skip_if_not_logged_in()
+
+  db <- connect("laminlabs/lamin-dev")
+
+  instance_settings <- db$get_settings()
+  expect_equal(instance_settings$name, "lamin-dev")
+})